requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
whitenoise==6.2.0
